backend/requirements.txt
asgiref==3.8.1
diskcache==5.6.3
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
Jinja2==3.1.6
llama_cpp_python==0.3.8
MarkupSafe==3.0.2
numpy==2.2.4
sqlparse==0.5.3
typing_extensions==4.13.1
